Agent finder: restore rich message bubble previews
